这个人很懒, 向你丢了一只随机皮卡丘(战斗力: 10101)!

+++ raptor-core/src/main/kotlin/as/leap/raptor/core/model/Header.kt
@@ -2,9 +2,9 @@ package `as`.leap.raptor.core.model
 
 data class Header(
     val fmt: FMT,
-    val csid: Short,
+    val csid: Int,
     val timestamp: Long = 0,
-    val streamId: Int,
+    val streamId: Long,
     val type: ChunkType,
     val length: Int?
 )
